rename

+++ easytcc-transaction/src/main/java/github/easytcc/transaction/LocalTransactionImpl.java
@@ -15,92 +15,101 @@ import github.easytcc.repository.factory.RepositoryFactory;
  * @author Fangfang.Xu
  *
  */
-public class LocalTransactionImpl extends AbstractTransaction implements LocalTransaction{
-	
+public class LocalTransactionImpl extends AbstractTransaction implements LocalTransaction {
+
 	private static final long serialVersionUID = 1L;
-	
+
 	static Logger logger = LoggerFactory.getLogger(LocalTransactionImpl.class);
-	
+
 	private String parentId;
-	
+
 	private static final int capcity = 5;
-	
+
 	private LocalParticipant[] localParticipants = new LocalParticipant[capcity];
 	/**
-	 * {@link participants} size
+	 * thread safe {@link participants} size
 	 */
 	private int size;
-	
+
 	private int scheduleRetryTimes;
-	
+
 	private RemoteTransaction remoteTransaction;
+
+	private boolean isRealtime;
 	/**
 	 * remoteTransaction will ignore if isRecovering was marked true
 	 */
 	private boolean isRecovering;
-	
-	public LocalTransactionImpl(String xid,String parentId,boolean realtime){
+
+	public LocalTransactionImpl(String xid, String parentId, boolean realtime) {
 		super(xid);
 		this.parentId = parentId;
-		if(realtime) {
+		this.isRealtime = realtime;
+		if (isRealtime) {
 			remoteTransaction = new RemoteTransaction(xid, this.getTransactionId());
-		}else {
-			remoteTransaction = new NonOpRemoteTransaction(xid, this.getTransactionId());
 		}
 	}
-	
+
 	@Override
 	public String getParentId() {
 		return parentId;
 	}
-	
+
 	@Override
 	public void addLocalParticipant(LocalParticipant localParticipant) {
-		if(size == localParticipants.length){
+		// ensure length
+		if (size == localParticipants.length) {
 			LocalParticipant[] dest = new LocalParticipant[capcity + localParticipants.length];
 			System.arraycopy(localParticipants, 0, dest, 0, localParticipants.length);
-			localParticipants = dest;			
+			localParticipants = dest;
 		}
+
 		localParticipants[size++] = localParticipant;
 	}
-	
+
 	@Override
 	public boolean commit() {
+		// only handle on realtime or recovering
+		if (!isRealtime && !isRecovering) {
+			return false;
+		}
+
 		boolean result = true;
-		if(remoteTransaction != null && !isRecovering){
+		if (remoteTransaction != null && !isRecovering) {
 			result = remoteTransaction.commit();
 		}
 		try {
 			List<LocalParticipant> participants = getLocalParticipants();
-			for(LocalParticipant participant:participants){
+			for (LocalParticipant participant : participants) {
 				participant.confirm();
 			}
-			try{
-				RepositoryFactory.getLocalTransactionRepository().deleteAndRemoveRetry(this);				
-			}catch (Exception e) {
-				//ignore
+			try {
+				RepositoryFactory.getLocalTransactionRepository().deleteAndRemoveRetry(this);
+			} catch (Exception e) {
+				// ignore
 			}
-		}catch (NeverBeRecoveryException e) {
-			//can not recovery,so remove from retrying, move associate xid to recoveryFailedXid collection
-			try{
+		} catch (NeverBeRecoveryException e) {
+			// can not recovery,so remove from retrying, move associate xid to
+			// recoveryFailedXid collection
+			try {
 				RepositoryFactory.getLocalTransactionRepository().addToRecoveryFailed(this);
-			}catch (Exception e1) {
-				//ignore
+			} catch (Exception e1) {
+				// ignore
 			}
-		}catch (Exception e) {
+		} catch (Exception e) {
 			result = false;
-			logger.error("commit local transaction failed",e);
-		}finally {
-			try{
+			logger.error("commit local transaction failed", e);
+		} finally {
+			try {
 				deleteAssociatedDownStreamApplications();
 			} catch (Exception e) {
 				// ignore
 			}
-			if(!result) {
-				try{
+			if (!result) {
+				try {
 					incrScheduleRetryTimes();
-				}catch (Exception e) {
-					//ignore
+				} catch (Exception e) {
+					// ignore
 				}
 			}
 		}
@@ -110,40 +119,41 @@ public class LocalTransactionImpl extends AbstractTransaction implements LocalTr
 	@Override
 	public boolean rollback() {
 		boolean result = true;
-		if(remoteTransaction != null && !isRecovering){
+		if (remoteTransaction != null && !isRecovering) {
 			result = remoteTransaction.rollback();
 		}
 		try {
 			List<LocalParticipant> participants = getLocalParticipants();
-			for(LocalParticipant participant:participants){
+			for (LocalParticipant participant : participants) {
 				participant.cancel();
-			} 
-			try{
-				RepositoryFactory.getLocalTransactionRepository().deleteAndRemoveRetry(this);	
-			}catch (Exception e) {
-				//ignore
 			}
-		}catch (NeverBeRecoveryException e) {
-			//can not recovery,so remove from retrying, move associate xid to recoveryFailedXid collection
-			try{
+			try {
+				RepositoryFactory.getLocalTransactionRepository().deleteAndRemoveRetry(this);
+			} catch (Exception e) {
+				// ignore
+			}
+		} catch (NeverBeRecoveryException e) {
+			// can not recovery,so remove from retrying, move associate xid to
+			// recoveryFailedXid collection
+			try {
 				RepositoryFactory.getLocalTransactionRepository().addToRecoveryFailed(this);
-			}catch (Exception e1) {
-				//ignore
+			} catch (Exception e1) {
+				// ignore
 			}
-		}catch (Exception e) {
+		} catch (Exception e) {
 			result = false;
-			logger.error("rollback local transaction failed",e);
-		}finally {
-			try{
+			logger.error("rollback local transaction failed", e);
+		} finally {
+			try {
 				deleteAssociatedDownStreamApplications();
 			} catch (Exception e) {
 				// ignore
 			}
-			if(!result) {
-				try{
+			if (!result) {
+				try {
 					incrScheduleRetryTimes();
-				}catch (Exception e) {
-					//ignore
+				} catch (Exception e) {
+					// ignore
 				}
 			}
 		}
@@ -152,14 +162,14 @@ public class LocalTransactionImpl extends AbstractTransaction implements LocalTr
 
 	private List<LocalParticipant> getLocalParticipants() {
 		List<LocalParticipant> list = new ArrayList<LocalParticipant>(localParticipants.length);
-		for(LocalParticipant participant:localParticipants){
-			if(participant!=null){
+		for (LocalParticipant participant : localParticipants) {
+			if (participant != null) {
 				list.add(participant);
 			}
 		}
 		return list;
 	}
-	
+
 	@Override
 	public void markedAsRecovering() {
 		isRecovering = true;
@@ -170,13 +180,14 @@ public class LocalTransactionImpl extends AbstractTransaction implements LocalTr
 	public void incrScheduleRetryTimes() {
 		scheduleRetryTimes++;
 		TccProperties tccProperties = SpringBeanFactory.getBean(TccProperties.class);
-		if(scheduleRetryTimes >= tccProperties.getRecoveryMaxRetrys()){
+		if (scheduleRetryTimes >= tccProperties.getRecoveryMaxRetrys()) {
 			RepositoryFactory.getLocalTransactionRepository().addToRecoveryFailed(this);
-			logger.warn("localTransaction.scheduleRetryTimes reached setted max times:{} but still failed so add to recovery failed collection,"
-					+ "serverName:{},localTransaction.id:{}",getScheduleRetryTimes(),
-					tccProperties.getApplication(),getTransactionId());
-		}else{
-			RepositoryFactory.getLocalTransactionRepository().save(this);			
+			logger.warn(
+					"localTransaction.scheduleRetryTimes reached setted max times:{} but still failed so add to recovery failed collection,"
+							+ "serverName:{},localTransaction.id:{}",
+					getScheduleRetryTimes(), tccProperties.getApplication(), getTransactionId());
+		} else {
+			RepositoryFactory.getLocalTransactionRepository().save(this);
 		}
 	}
 
@@ -184,9 +195,9 @@ public class LocalTransactionImpl extends AbstractTransaction implements LocalTr
 	public int getScheduleRetryTimes() {
 		return scheduleRetryTimes;
 	}
-	
+
 	private void deleteAssociatedDownStreamApplications() {
-		if(getParentId() != null) {
+		if (getParentId() != null) {
 			RepositoryFactory.getTransactionDownstreamRepository()
 					.deleteAssociatedDownStreamApplications(getParentId());
 			this.parentId = null;
